+++ src-ui/app/services/brightness-control/display-brightness-drivers/display-brightness-control-driver.ts
@@ -1,8 +1,0 @@
-import { Observable } from 'rxjs';
-
-export abstract class DisplayBrightnessControlDriver {
-  abstract getBrightnessPercentage(): Promise<number>;
-  abstract setBrightnessPercentage(percentage: number): Promise<void>;
-  abstract getBrightnessBounds(): Promise<[number, number]>;
-  abstract isAvailable(): Observable<boolean>;
-}