+++ src-ui/app/services/brightness-control/software-brightness-control.service.ts
@@ -7,16 +7,16 @@ import { invoke } from '@tauri-apps/api';
 import { SET_BRIGHTNESS_OPTIONS_DEFAULTS, SetBrightnessOptions } from './brightness-control-models';
 import { listen } from '@tauri-apps/api/event';
 
-export const DEFAULT_IMAGE_BRIGHTNESS_GAMMA = 0.55;
+export const DEFAULT_SOFTWARE_BRIGHTNESS_GAMMA = 0.55;
 
 @Injectable({
   providedIn: 'root',
 })
-export class ImageBrightnessControlService {
+export class SoftwareBrightnessControlService {
   private _brightness: BehaviorSubject<number> = new BehaviorSubject<number>(100);
   private _activeTransition = new BehaviorSubject<BrightnessTransitionTask | undefined>(undefined);
   public readonly activeTransition = this._activeTransition.asObservable();
-  private _perceivedBrightnessAdjustmentGamma: number | null = DEFAULT_IMAGE_BRIGHTNESS_GAMMA;
+  private _perceivedBrightnessAdjustmentGamma: number | null = DEFAULT_SOFTWARE_BRIGHTNESS_GAMMA;
 
   get brightness(): number {
     return this._brightness.value;
@@ -28,7 +28,7 @@ export class ImageBrightnessControlService {
 
   public set perceivedBrightnessAdjustmentGamma(value: number | null) {
     this._perceivedBrightnessAdjustmentGamma = value;
-    this.setImageBrightness(this.brightness);
+    this.setSoftwareBrightness(this.brightness);
   }
 
   public get perceivedBrightnessAdjustmentGamma() {
@@ -36,13 +36,13 @@ export class ImageBrightnessControlService {
   }
 
   async init() {
-    await this.setImageBrightness(this.brightness);
-    await listen<number>('setImageBrightness', async (event) => {
+    await this.setSoftwareBrightness(this.brightness);
+    await listen<number>('setSoftwareBrightness', async (event) => {
       await this.setBrightness(event.payload, { cancelActiveTransition: true });
     });
   }
 
-  private async setImageBrightness(brightness: number) {
+  private async setSoftwareBrightness(brightness: number) {
     await invoke('openvr_set_image_brightness', {
       brightness: brightness / 100,
       perceivedBrightnessAdjustmentGamma: this._perceivedBrightnessAdjustmentGamma ?? null,
@@ -62,7 +62,7 @@ export class ImageBrightnessControlService {
     }
     this._activeTransition.value?.cancel();
     const transition = new BrightnessTransitionTask(
-      'IMAGE',
+      'SOFTWARE',
       this.setBrightness.bind(this),
       async () => this.brightness,
       async () => [0, 100],
@@ -79,7 +79,9 @@ export class ImageBrightnessControlService {
         this._activeTransition.next(undefined);
     });
     if (opt.logReason) {
-      info(`[BrightnessControl] Starting image brightness transition (Reason: ${opt.logReason})`);
+      info(
+        `[BrightnessControl] Starting software brightness transition (Reason: ${opt.logReason})`
+      );
     }
     this._activeTransition.next(transition);
     transition.start();
@@ -101,10 +103,10 @@ export class ImageBrightnessControlService {
     if (opt.cancelActiveTransition) this.cancelActiveTransition();
     if (percentage == this.brightness) return;
     this._brightness.next(percentage);
-    await this.setImageBrightness(percentage);
+    await this.setSoftwareBrightness(percentage);
     if (opt.logReason) {
       await info(
-        `[BrightnessControl] Set image brightness to ${percentage}% (Reason: ${opt.logReason})`
+        `[BrightnessControl] Set software brightness to ${percentage}% (Reason: ${opt.logReason})`
       );
     }
   }